+++ microlens-th/CHANGELOG.md
@@ -1,3 +1,7 @@
+# next
+
+* `createClass` was removed because it doesn't seem to be useful without `lensClass`.
+
 # 0.1.2.0
 
 * Package now compiles with `-O2` and other optimisations by default.